feat(profiles): Deye/Sunsynk SUN-xK-SG hybrid, and what two sources would not agree on
The first profile for a vendor other than the one the driver used to be named
after. Every mapped row is corroborated by two INDEPENDENT implementations read
directly: kellerza/sunsynk's single_phase.py and kbialek/deye-inverter-mqtt's
deye_sensors_deye_hybrid.py. Holding registers throughout, 9600 8N1.

Two traps caught while cross-checking, both recorded in the profile because both
would have quietly corrupted the map:

  - source A's factor sign is not a negation. Sensor(190, ..., -1) means "read
    this register as SIGNED": reg_to_value() does unpack_value(signed=factor<0)
    and multiplies by abs(factor). Read as a negation it inverts every signed row;
  - solarman's deye_hybrid.yaml is NOT a third source. It is byte-identical to
    source B's own generated ha_definitions file (same MD5). Counting it would
    have turned one implementation into two and defeated the cross-check.

And one that decided four rows: source A assigns 182/183/184/186 TWICE -- its
AUX/generator block claims 181-186 while its battery and PV blocks claim the same
registers for battery temperature, voltage, SoC and PV1 power. Source B has no AUX
block and assigns exactly the battery/PV meanings, so an independent implementation
breaks the tie, and breaks it against what looks like a three-phase map pasted into
the single-phase file.

WHAT IS DELIBERATELY ABSENT matters as much as what is here, and a test pins each
one as absent so it cannot drift back in without a bench reading:

  - battery.power and battery.current: both sources agree on address, scale and
    signedness, and NEITHER states which direction positive means. An inverted
    battery graph is worse than no battery graph, because nothing about it looks
    wrong;
  - ac.frequency: source A maps grid frequency to 79, source B has no register 79
    at all, and the only frequency both carry (192) is called the LOAD frequency by
    one and the AC frequency by the other. On a hybrid those differ;
  - inverter.temperature: both agree 90 and 91 are temperatures with the same
    decoding, and disagree about which is the heatsink;
  - dc.power.total, grid import/export, and the third string: no agreed register,
    no way to split one signed bidirectional value into two rails without
    arithmetic, and one source respectively.

Battery temperature IS mapped, and it is the row that justified `offset`: both
sources independently decode register 182 as raw * 0.1 - 100.

The declared setpoint (holding 244, load handling) is dormant twice over, and the
second reason is the one a flipped `verified` flag would not fix: both sources write
this family with FC16 -- source A says so outright, source B's frame builder emits
01 10 unconditionally, neither ever sends FC06. So it is declared
function = "write_multiple", which this firmware's write path refuses by design.
Enabling it needs FC16 support, not a flag.

Also here, from the three-agent review of this branch:

  - test_discovery_signature_changes_when_a_mode_is_renumbered narrowed a uint64_t
    signature into a uint32_t and then compared it against full-width values, so the
    assertions were true whatever the signature did. A test that could not fail,
    guarding the invariant most worth guarding. Fixed, and confirmed by mutation:
    it now fails when the fold-in of option values is removed, and passes when it
    is restored;
  - the profile generator accepted `nan`/`inf` for scale and offset (TOML has them
    as literals) and emitted `nan.0` / bare `inf` -- an opaque compiler error
    instead of a diagnostic naming the file;
  - it accepted control characters in any display name, manufacturer or option
    label; a TOML "\n" escape becomes a real newline inside a C++ string literal;
  - two DIFFERENT profile ids could collapse to one C++ symbol (a_b and a__b both
    give AB), passing the duplicate-id check and colliding in generated code;
  - check_layering.sh's LEGACY-CONFIG-ID exemption was a bare per-line match, so a
    second marker anywhere could hide a real brand-name leak -- the comment claimed
    it "cannot quietly widen", which was not true as implemented. Now bounded to
    exactly one occurrence;
  - schema.md's `function` row still invited FC16-for-a-single-register, which the
    driver refuses -- a contradiction with the write-path doc added two commits ago
    -- and its `unit`/`step` rows did not say they are refused on a mode row.

Every generator rejection above verified by feeding it a profile that trips it.

Documented, not silently shipped: the rename changes the device id of every
non-primary device from growatt_modbus-N to modbus_profile-N, so Home Assistant
retires those entities and their recorder history does not carry over. Primary
devices are keyed on the bridge id and are unaffected. Upgrade note added to the
multi-device instructions.

904 native tests pass; layering checks pass; all board envs build.

+++ test/test_mqtt/test_main.cpp
@@ -779,7 +779,11 @@ static void test_discovery_signature_changes_when_a_mode_is_renumbered() {
     static constexpr EnumOption kRenamed[] = {
         {0, "Self-consumption"}, {2, "Compulsory"}, {3, "External EMS"}};
 
-    const uint32_t base = discoverySignature(modeState(kTestModes, 3));
+    // uint64_t, matching what discoverySignature returns. Narrowing this to uint32_t threw away
+    // the top half of the hash and then compared it against full 64-bit values, so the
+    // assertions below were true whatever the signature did -- a test that could not fail,
+    // guarding the one invariant most worth guarding.
+    const uint64_t base = discoverySignature(modeState(kTestModes, 3));
     TEST_ASSERT_TRUE(base != discoverySignature(modeState(kRenumbered, 3)));
     TEST_ASSERT_TRUE(base != discoverySignature(modeState(kRenamed, 3)));
     // And dropping one is a change too, not merely a shorter list that hashes the same.
